fix: fix LobbyNav links to link to correct paths
Using usePathName as the base of the URL path and concatenating it with
another string path creates incorrect URL navigation. For example,
if the user click the communities nav link, they will correctly directed
to 'lobby/communities' page, but, once the user is in directed to
the communities page and click the communities link again, they will
get directed to 'lobby/communities/communities' - this is due to
concatenating path string with the base path from usePathName.