fix: use community id as the folder's name
Using the community name as the folder name to save the community's
banner and avatar causes inconsistent folder structures on Supabase.

For example, Supabase fails to create the folder when the community's
name contains certain characters, this causes the images to be saved
inside the root bucket folder instead of a separate partitioned
folder.

Fix this issue by using the community id as the folder name to prevent
bugs that cause improper folder structures and improve folder naming
and structure consistency.