fix(queries):fix getUserCommunities
The getUserCommunities query returns incorrect data. The previous query
improperly uses prisma's every filter which means it returns
communities where every participant's userId is equal to the userId
argument.

It also doesn't account for the participant's status and treats a
PENDING user as an ACTIVE member of the community. This issue is fixed
by adding a filter on the status field. This ensures that the
getUserCommunities only returns communities where the userId is an
ACTIVE participant.