fix: improve explore communities ISSUE 0003
The communities are sorted by the following ORDER BY clause:
 - Order by the most participants
 - Order by newest createdAt data
 - Order id descendingly

This query uses participantsCount as the main order condition, and it is
stabilised by using createdAt and id ORDER BY and filters.
This reduces the probability of fetching duplicates, but it hasn't been
tested on non-static data. This means this query might fetch the same
communities in the production server due to participantsCount changing
on each fetch.

The current WHERE clause ensures that no duplicates are fetched:

- Select communities with fewer participants compared to the cursor's
participants.

- Select communities with equal participants to the cursor's
participants, but only selecting communities that were created before
the cursor's createdAt or communities with the same creation date, but
with id smaller than the cursor's id.

This query doesn't return all the expected communities that match the
search name param, but it does improve pagination stability when
fetching static communities data. While this query returns patchy data,
it is the current best solution when ordering and filtering using a
volatile field.

Fixes #3