test(controllers): stabillse community_get_explore
Use the middle community to test the returned cursor to improve test
stability due to the changed participantsCount on the seeded data. This
is due to the populate script populating more participantsCount in the
middle of the communities array.

To stabilise the test, instead of using community102 we'll pick the
community that is positioned in the middle to test the returned
cursor. This way, we'll account for probable gaps that happen in the
search communities query and future seed data changes.