fix: fix friendNONE query in friend request test
Fix logic error of using 'AND' instead of 'OR' to filter out 'seeduser2'
and 'seeduser1' from being selected for the friend test. The 'AND'
statement is used incorrectly and is not the expected behaviour when
selecting user for friend requests, which causes it to interfere with
other tests when seeduser2 is selected.